requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
dotenv==0.9.9
fonttools==4.61.1
greenlet==3.3.0
idna==3.11
joblib==1.5.3
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.8
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
psycopg==3.3.2
psycopg-binary==3.3.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
SQLAlchemy==2.0.45
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
